Rename html.py to html_utils.py for clarity and to please ruff
